requirements.txt
pandas==1.4.4
numpy==1.23.3
scipy==1.9.1
bokeh==2.4.3
plotly==5.10.0
zipfile36
opencv-python==4.6.0.66
matplotlib==3.6.0
subprocess.run
gunicorn==20.1.0
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-loading-spinners==1.0.0
dash-table==5.0.0
dash-uploader==0.6.0
packaging==21.3
torchaudio==0.12.1
torchvision==0.13.1
waitress==2.1.2
